refactoring related indicator table
