+++ src/Module/DataElement/index.js
@@ -5,7 +5,7 @@ import AnalyticsDetails from "./Components/AnalyticsDetails";
 
 import AccessibilityAndSharing from "../../Shared/Componets/AccesibilityAndSharing";
 import React from 'react'
-import RelatedIndicator from "../../Shared/Componets/RelatedIndicator";
+import RelatedIndicator from "../../Shared/Componets/RelatedIndicatorTable";
 import Introduction from "./Components/introduction/introduction";
 
 
@@ -17,7 +17,10 @@ export default function DataElementPage(props){
             <DataSource id={id}  />
             <Facts  id={id}  />
             <AnalyticsDetails id={id}/>
+
+            <h3>{i18n.t("Related Indicators")} </h3>
             <RelatedIndicator id={id} resourceType={"Data Element"}  />
+
             <AccessibilityAndSharing id={id} resourceType={"dataElements"} />
 
     </div>
